pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "linkedge"
version = "0.1.0"
description = "Generate LinkedIn profile banner images with QR codes"
requires-python = ">=3.11"
dependencies = [
    "pillow>=10.1",
    "qrcode[pil]",
    "typer[all]",
    "rich",
    "pydantic",
]

[project.scripts]
linkedge = "linkedge.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/linkedge"]

[dependency-groups]
dev = [
    "ruff",
    "mypy",
    "pytest",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "RUF"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
